Use `u8` instead of `usize` for array map (#341)

+++ compiler+runtime/include/cpp/jank/runtime/detail/native_array_map.hpp
@@ -20,7 +20,7 @@ namespace jank::runtime::detail
     /* Array maps are fast only for a small number of keys. Clojure JVM uses a threshold of 8
      * k/v pairs, thus 16 elements. We follow the same. */
     /* TODO: Benchmark difference thresholds. */
-    static constexpr usize max_size{ 8 };
+    static constexpr u8 max_size{ 8 };
 
     native_array_map() = default;
     native_array_map(native_array_map const &s) = default;
@@ -53,7 +53,7 @@ namespace jank::runtime::detail
       using pointer = value_type *;
       using reference = value_type &;
 
-      iterator(object_ref const *data, usize index);
+      iterator(object_ref const *data, u8 index);
       iterator(iterator const &) = default;
       iterator(iterator &&) noexcept = default;
 
@@ -68,7 +68,7 @@ namespace jank::runtime::detail
       iterator &operator=(iterator const &rhs);
 
       object_ref const *data{};
-      usize index{};
+      u8 index{};
     };
 
     using const_iterator = iterator;
@@ -76,18 +76,18 @@ namespace jank::runtime::detail
     const_iterator begin() const;
     const_iterator end() const;
 
-    void reserve(usize const size);
+    void reserve(u8 const size);
 
-    usize capacity() const;
-    usize size() const;
+    u8 capacity() const;
+    u8 size() const;
 
     bool empty() const;
 
     native_array_map clone() const;
 
     object_ref *data{};
-    usize cap{};
-    usize length{};
+    u8 cap{};
+    u8 length{};
     mutable uhash hash{};
   };
 }
